Fix published property updates on wrong thread. #153
I'm new to Swift structured concurrency but it looks like some of the async functions called from views are not using @MainActor so their updates come in on the wrong thread when the requests complete. Adding @MainActor resolves it, and I don't think it is pushing all the work onto the main queue as the network requests have their own actor/queue I presume.